+++ node/resolvers/Routes/index.ts
@@ -271,10 +271,34 @@ export const Routes = {
     ) {
       const [address] = costCenterResponse.data.getCostCenterById.addresses
 
+      const marketingTagsResponse: any = await graphqlServer
+        .query(
+          QUERIES.getMarketingTags,
+          {
+            costId: user.costId,
+          },
+          {
+            persistedQuery: {
+              provider: 'vtex.b2b-organizations-graphql@0.x',
+              sender: 'vtex.storefront-permissions@1.x',
+            },
+          }
+        )
+        .catch((error) => {
+          logger.error({
+            error,
+            message: 'setProfile.getMarketingTags',
+          })
+        })
+
+      const marketingTags: any =
+        marketingTagsResponse?.data?.getMarketingTags?.tags
+
       promises.push(
         checkout
           .updateOrderFormMarketingData(orderFormId, {
             attachmentId: 'marketingData',
+            marketingTags: marketingTags || [],
             utmCampaign: user.orgId,
             utmMedium: user.costId,
           })