+++ node/resolvers/Routes/utils/index.ts
@@ -35,6 +35,13 @@ export const QUERIES = {
         businessDocument
       }
     }`,
+  getMarketingTags: `
+    query ($costId: ID!) {
+      getMarketingTags(costId: $costId) @context(provider: "vtex.b2b-organizations-graphql") {
+        tags
+      }
+    }
+  `,
   getOrganizationById: `query Organization($id: ID!){
       getOrganizationById(id: $id) @context(provider: "vtex.b2b-organizations-graphql"){
         name